// src/shared-report-level/enums/function-parameters.enum.ts
export enum FunctionParametersEnum {
    Attributes = 'Attributes',
    SortOrder = 'SortOrder',
    Offset = 'Offset',
    GroupCount = 'GroupCount',
    PercentileValue = 'PercentileValue',
    ContributionPct = 'ContributionPct',
    SortOn = 'SortOn',
    WindowStart = 'WindowStart',
    WindowEnd = 'WindowEnd'
}